telegram-buddy-ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.51.0
anthropic==0.7.0
python-multipart==0.0.6
python-telegram-bot==20.7
